refactor: split doc provider modules
