+++ packages/store/src/persistence/doc/index.ts
@@ -1,9 +1,8 @@
 import type * as Y from 'yjs';
-// @ts-ignore
-import { WebrtcProvider } from 'y-webrtc';
-import { IndexedDBPersistence } from './indexeddb.js';
 import type { Awareness } from 'y-protocols/awareness';
-import { isWeb } from '@blocksuite/global/utils';
+
+export { DebugDocProvider } from './debug-provider.js';
+export { IndexedDBDocProvider } from './idb-provider.js';
 
 /**
  * Different examples of providers could include webrtc sync,
@@ -26,58 +25,4 @@ export interface DocProviderConstructor {
     doc: Y.Doc,
     options?: { awareness?: Awareness }
   ): DocProvider;
-}
-
-// When using playground from blocksuite repo, this comes from "serve" script in "@blocksuite/store" package.
-// We use our own sync server because a local service for sync makes everything much faster for dev.
-const LOCAL_SIGNALING = ['ws://localhost:4444'];
-
-const DEFAULT_SIGNALING = [
-  // Default config from yjs (but these are kinda slow by comparison to self host sync ~100ms-+1000ms latency observed).
-  // This slowness is also avoided in order to improve test reliability in CI.
-  'wss://y-webrtc-signaling-us.herokuapp.com',
-  'wss://y-webrtc-signaling-eu.herokuapp.com',
-];
-
-const isLocalhost =
-  isWeb &&
-  (window.location.hostname === 'localhost' ||
-    window.location.hostname === '127.0.0.1');
-const signaling = isLocalhost ? LOCAL_SIGNALING : DEFAULT_SIGNALING;
-
-export class DebugDocProvider extends WebrtcProvider implements DocProvider {
-  constructor(room: string, doc: Y.Doc, options?: { awareness?: Awareness }) {
-    super(room, doc, {
-      awareness: options?.awareness,
-      signaling,
-    });
-  }
-
-  public clearData() {
-    // Do nothing for now
-    return Promise.resolve();
-  }
-}
-
-export class IndexedDBDocProvider
-  extends IndexedDBPersistence
-  implements DocProvider
-{
-  constructor(room: string, doc: Y.Doc) {
-    super(room, doc);
-  }
-
-  // Consider whether "connect" and "disconnect" are good to put on the SyncProvider
-  connect() {
-    // not necessary as it will be set up in indexeddb persistence
-  }
-
-  disconnect() {
-    // not necessary as it will be set up in indexeddb persistence
-  }
-
-  public clearData() {
-    // Do nothing for now
-    return Promise.resolve();
-  }
 }
